Implementing /number-of-files endpoint
This endpoint allows getting the total number of files
in the storage service. It simply reads an atomic long
variable that keeps track of the total number of files
stored.

Such an approach should not introduce too much overhead
since the only bottleneck for the file access operations
remains the update of this variable (everything else can
still happen in parallel).