Add customizable imbrication level for storage
Reading the maximum imbrication level for storage service
for application.properties file. This allows configuring
the max depth of the hierarchy of directories in which
files are stored. This could be a handful parameter that
could be used to reduce access time for files. A value of
3 should be a balanced option between the overhead introduced
by a larger number of folders (from the OS's filesystem point
of view) and the reduction in access time to files that we
get if we reduce the number of files stored in a directory
(for most OSs, accessing a file means passing through linked
lists that stores metadata about the content of each directory
so having less files in folders means having shorter linked
lists to pass meaning shorter access times).